fix(dashboard): address 2nd-pass review findings
- governance-actions: read resource state directly after load() instead
  of relying on catch block (load() swallows errors internally)
- agent-profile: only reset() when agent name changes, preventing
  UI loading flash on same-agent refresh
- logs: replace useSignalEffect with useEffect to prevent re-entrance
  loop from reset() writing to tracked signal
- Extract StatCard to common/stat-card.ts, removing duplication between
  feature-health.ts and harness-health-sections.ts
