+++ dashboard/src/components/governance-actions.ts
@@ -50,20 +50,21 @@ export async function selectDecision(item: GovernanceDecisionItem) {
 
 export async function refreshGovernance() {
   governanceError.value = ''
-  try {
-    await governanceResource.load(async () => {
-      return await fetchDashboardGovernance()
-    })
-    const data = governanceData.value
-    if (data) {
-      const items = filteredItemsByFilter(governanceFilter.value, data.items ?? [])
-      const current = selectedDecisionKey.value
-      const next = items.find(item => itemKey(item) === current) ?? items[0] ?? null
-      selectedDecisionKey.value = next ? itemKey(next) : null
-      await loadDecisionDetail(next)
-    }
-  } catch (err) {
-    governanceError.value = err instanceof Error ? err.message : '거버넌스 상태를 불러오지 못했습니다'
+  await governanceResource.load(async () => {
+    return await fetchDashboardGovernance()
+  })
+  const s = governanceResource.state.value
+  if (s.status === 'error') {
+    governanceError.value = s.message
+    return
+  }
+  if (s.status === 'loaded') {
+    const data = s.data
+    const items = filteredItemsByFilter(governanceFilter.value, data.items ?? [])
+    const current = selectedDecisionKey.value
+    const next = items.find(item => itemKey(item) === current) ?? items[0] ?? null
+    selectedDecisionKey.value = next ? itemKey(next) : null
+    await loadDecisionDetail(next)
   }
 }
