+++ dashboard/src/components/harness-health-sections.ts
@@ -152,15 +152,7 @@ export function StatusPill({ status }: { status: RailStatus }) {
   `
 }
 
-export function StatCard({ label, value, sub }: { label: string; value: string | number; sub?: string }) {
-  return html`
-    <div class="rounded-lg border border-[var(--white-8)] bg-[var(--white-4)] p-3 text-center">
-      <div class="text-2xl font-bold text-[var(--accent)]">${value}</div>
-      <div class="mt-1 text-xs text-[var(--text-muted)]">${label}</div>
-      ${sub ? html`<div class="mt-0.5 text-xs text-[var(--text-dim)]">${sub}</div>` : null}
-    </div>
-  `
-}
+export { StatCard } from './common/stat-card'
 
 export function EmptySignal({ text }: { text: string }) {
   return html`
